feat(base): allow for 0-approval flow
This commit removes the approval check, from happening in the
'/approve' endpoint.
Instead, all necessary functionality for an approved AccessRequest
(GrantCode issuance, tying it with the AccessRequest, storing the
mapping), is happening every time an AccessRequest is fetched.

This way, even if the '/approve' endpoint is never hit, an
AccessRequest can be claimable (if the 'required_approvals' is 0).