feat: add adhoc revocation to `policy-gate`
This rewrite contains all previous features (of base and mock),
and also adds the need for the plugin to interact with
Vault/OpenBao directly, through an AppRole.

This AppRole has to be created individually, and it is taken care
of in the [`gateplane-io/terraform-gateplane-policy-gate`](https://github.com/gateplane-io/terraform-gateplane-policy-gate) module.

The way it works is now dfferent, as it adds and removes policies
directly to and from the Entity Policies of the requesting entity,
no needing a different (elevated) token to be claimed, and not
needing re-authentication of the requestor.